refactor(dup): extract event dispatching logic to Astarte.Events.TriggersHandler
- Move dispatching logic from Astarte.DataUpdaterPlant.TriggersHandler into astarte_events
- Move make_simple_event function and related to astarte_events
