+++ apps/astarte_data_updater_plant/lib/astarte_data_updater_plant/triggers_handler.ex
@@ -17,19 +17,15 @@
 #
 
 defmodule Astarte.DataUpdaterPlant.TriggersHandler do
-  import Bitwise, only: [<<<: 2]
   require Logger
-  alias Astarte.Events.AMQPEventsProducer
+  alias Astarte.Events.TriggersHandler
+
   alias Astarte.DataUpdaterPlant.Config
 
   @moduledoc """
   This module handles the triggers by generating the events requested
   by the Trigger targets
   """
-
-  @max_backoff_exponent 8
-  @max_rand trunc(:math.pow(2, 32) - 1)
-
   alias Astarte.Core.Triggers.SimpleEvents.{
     DeviceConnectedEvent,
     DeviceDisconnectedEvent,
@@ -42,7 +38,6 @@ defmodule Astarte.DataUpdaterPlant.TriggersHandler do
     InterfaceVersion,
     PathCreatedEvent,
     PathRemovedEvent,
-    SimpleEvent,
     ValueChangeAppliedEvent,
     ValueChangeEvent
   }
@@ -66,15 +61,14 @@ defmodule Astarte.DataUpdaterPlant.TriggersHandler do
 
   def device_connected(target, realm, device_id, ip_address, timestamp, policy) do
     %DeviceConnectedEvent{device_ip_address: ip_address}
-    |> make_simple_event(
+    |> dispatch_event_with_telemetry(
       :device_connected_event,
-      target.simple_trigger_id,
-      target.parent_trigger_id,
+      target,
       realm,
       device_id,
-      timestamp
+      timestamp,
+      policy
     )
-    |> dispatch_event(target, policy)
   end
 
   def device_disconnected(targets, realm, device_id, timestamp) when is_list(targets) do
@@ -85,15 +79,14 @@ defmodule Astarte.DataUpdaterPlant.TriggersHandler do
 
   def device_disconnected(target, realm, device_id, timestamp, policy) do
     %DeviceDisconnectedEvent{}
-    |> make_simple_event(
+    |> dispatch_event_with_telemetry(
       :device_disconnected_event,
-      target.simple_trigger_id,
-      target.parent_trigger_id,
+      target,
       realm,
       device_id,
-      timestamp
+      timestamp,
+      policy
     )
-    |> dispatch_event(target, policy)
   end
 
   def device_error(targets, realm, device_id, error_name, error_metadata, timestamp)
@@ -115,15 +108,14 @@ defmodule Astarte.DataUpdaterPlant.TriggersHandler do
     metadata_kw = Enum.into(error_metadata, [])
 
     %DeviceErrorEvent{error_name: error_name, metadata: metadata_kw}
-    |> make_simple_event(
+    |> dispatch_event_with_telemetry(
       :device_error_event,
-      target.simple_trigger_id,
-      target.parent_trigger_id,
+      target,
       realm,
       device_id,
-      timestamp
+      timestamp,
+      policy
     )
-    |> dispatch_event(target, policy)
   end
 
   def incoming_data(targets, realm, device_id, interface, path, bson_value, timestamp)
@@ -145,15 +137,14 @@ defmodule Astarte.DataUpdaterPlant.TriggersHandler do
         policy
       ) do
     %IncomingDataEvent{interface: interface, path: path, bson_value: bson_value}
-    |> make_simple_event(
+    |> dispatch_event_with_telemetry(
       :incoming_data_event,
-      target.simple_trigger_id,
-      target.parent_trigger_id,
+      target,
       realm,
       device_id,
-      timestamp
+      timestamp,
+      policy
     )
-    |> dispatch_event(target, policy)
   end
 
   def incoming_introspection(targets, realm, device_id, introspection, timestamp)
@@ -180,15 +171,14 @@ defmodule Astarte.DataUpdaterPlant.TriggersHandler do
       end
 
     incoming_introspection_event
-    |> make_simple_event(
+    |> dispatch_event_with_telemetry(
       :incoming_introspection_event,
-      target.simple_trigger_id,
-      target.parent_trigger_id,
+      target,
       realm,
       device_id,
-      timestamp
+      timestamp,
+      policy
     )
-    |> dispatch_event(target, policy)
   end
 
   def interface_added(
@@ -240,15 +230,14 @@ defmodule Astarte.DataUpdaterPlant.TriggersHandler do
       major_version: major_version,
       minor_version: minor_version
     }
-    |> make_simple_event(
+    |> dispatch_event_with_telemetry(
       :interface_added_event,
-      target.simple_trigger_id,
-      target.parent_trigger_id,
+      target,
       realm,
       device_id,
-      timestamp
+      timestamp,
+      policy
     )
-    |> dispatch_event(target, policy)
   end
 
   def interface_minor_updated(
@@ -294,15 +283,14 @@ defmodule Astarte.DataUpdaterPlant.TriggersHandler do
       old_minor_version: old_minor,
       new_minor_version: new_minor
     }
-    |> make_simple_event(
+    |> dispatch_event_with_telemetry(
       :interface_minor_updated_event,
-      target.simple_trigger_id,
-      target.parent_trigger_id,
+      target,
       realm,
       device_id,
-      timestamp
+      timestamp,
+      policy
     )
-    |> dispatch_event(target, policy)
   end
 
   def interface_removed(targets, realm, device_id, interface, major_version, timestamp)
@@ -323,15 +311,14 @@ defmodule Astarte.DataUpdaterPlant.TriggersHandler do
         policy
       ) do
     %InterfaceRemovedEvent{interface: interface, major_version: major_version}
-    |> make_simple_event(
+    |> dispatch_event_with_telemetry(
       :interface_removed_event,
-      target.simple_trigger_id,
-      target.parent_trigger_id,
+      target,
       realm,
       device_id,
-      timestamp
+      timestamp,
+      policy
     )
-    |> dispatch_event(target, policy)
   end
 
   def path_created(targets, realm, device_id, interface, path, bson_value, timestamp)
@@ -353,15 +340,14 @@ defmodule Astarte.DataUpdaterPlant.TriggersHandler do
         policy
       ) do
     %PathCreatedEvent{interface: interface, path: path, bson_value: bson_value}
-    |> make_simple_event(
+    |> dispatch_event_with_telemetry(
       :path_created_event,
-      target.simple_trigger_id,
-      target.parent_trigger_id,
+      target,
       realm,
       device_id,
-      timestamp
+      timestamp,
+      policy
     )
-    |> dispatch_event(target, policy)
   end
 
   def path_removed(targets, realm, device_id, interface, path, timestamp) when is_list(targets) do
@@ -372,15 +358,14 @@ defmodule Astarte.DataUpdaterPlant.TriggersHandler do
 
   def path_removed(target, realm, device_id, interface, path, timestamp, policy) do
     %PathRemovedEvent{interface: interface, path: path}
-    |> make_simple_event(
+    |> dispatch_event_with_telemetry(
       :path_removed_event,
-      target.simple_trigger_id,
-      target.parent_trigger_id,
+      target,
       realm,
       device_id,
-      timestamp
+      timestamp,
+      policy
     )
-    |> dispatch_event(target, policy)
   end
 
   def value_change(
@@ -427,15 +412,14 @@ defmodule Astarte.DataUpdaterPlant.TriggersHandler do
       old_bson_value: old_bson_value,
       new_bson_value: new_bson_value
     }
-    |> make_simple_event(
+    |> dispatch_event_with_telemetry(
       :value_change_event,
-      target.simple_trigger_id,
-      target.parent_trigger_id,
+      target,
       realm,
       device_id,
-      timestamp
+      timestamp,
+      policy
     )
-    |> dispatch_event(target, policy)
   end
 
   def value_change_applied(
@@ -481,34 +465,14 @@ defmodule Astarte.DataUpdaterPlant.TriggersHandler do
       old_bson_value: old_bson_value,
       new_bson_value: new_bson_value
     }
-    |> make_simple_event(
+    |> dispatch_event_with_telemetry(
       :value_change_applied_event,
-      target.simple_trigger_id,
-      target.parent_trigger_id,
+      target,
       realm,
       device_id,
-      timestamp
+      timestamp,
+      policy
     )
-    |> dispatch_event(target, policy)
-  end
-
-  defp make_simple_event(
-         event,
-         event_type,
-         simple_trigger_id,
-         parent_trigger_id,
-         realm,
-         device_id,
-         timestamp
-       ) do
-    %SimpleEvent{
-      simple_trigger_id: simple_trigger_id,
-      parent_trigger_id: parent_trigger_id,
-      realm: realm,
-      device_id: device_id,
-      timestamp: timestamp,
-      event: {event_type, event}
-    }
   end
 
   defp introspection_string_to_introspection_proto_map!(introspection_string) do
@@ -536,178 +500,35 @@ defmodule Astarte.DataUpdaterPlant.TriggersHandler do
     end
   end
 
-  defp wait_backoff_and_publish(
-         :ok,
-         _realm,
-         _retry,
-         _exchange,
-         _routing_key,
-         _payload,
-         _opts
-       ) do
-    :ok
-  end
-
-  defp wait_backoff_and_publish(
-         {:error, reason},
+  defp dispatch_event_with_telemetry(
+         event,
+         event_type,
+         target,
          realm,
-         retry,
-         exchange,
-         routing_key,
-         payload,
-         opts
+         device_id,
+         timestamp,
+         policy
        ) do
-    Logger.warning(
-      "Failed publish on events exchange with #{routing_key}. Reason: #{inspect(reason)}"
-    )
-
-    retry
-    |> compute_backoff_time()
-    |> :timer.sleep()
-
-    next_retry =
-      if retry <= @max_backoff_exponent do
-        retry + 1
-      else
-        retry
-      end
-
-    publish(realm, exchange, routing_key, payload, opts)
-    |> wait_backoff_and_publish(realm, next_retry, exchange, routing_key, payload, opts)
-  end
-
-  defp wait_ok_publish(realm, exchange, routing_key, payload, opts) do
-    publish(realm, exchange, routing_key, payload, opts)
-    |> wait_backoff_and_publish(realm, 1, nil, routing_key, payload, opts)
-  end
-
-  defp publish(_realm, nil, routing_key, payload, opts) do
-    Config.events_exchange_name!()
-    |> AMQPEventsProducer.publish(routing_key, payload, opts)
-  end
-
-  defp publish(realm, exchange, routing_key, payload, opts) do
-    Astarte.Events.AMQPTriggers.publish(realm, exchange, routing_key, payload, opts)
-  end
-
-  defp dispatch_event(
-         simple_event = %SimpleEvent{},
-         %AMQPTriggerTarget{
-           exchange: target_exchange,
-           routing_key: routing_key,
-           static_headers: static_headers,
-           message_expiration_ms: message_expiration_ms,
-           message_priority: message_priority,
-           message_persistent: message_persistent
-         },
-         policy_name
-       ) do
-    {event_type, _event_struct} = simple_event.event
-
-    simple_trigger_id_str =
-      simple_event.simple_trigger_id
-      |> :uuid.uuid_to_string()
-      |> to_string()
-
-    parent_trigger_id_str =
-      simple_event.parent_trigger_id
-      |> :uuid.uuid_to_string()
-      |> to_string()
-
-    headers = [
-      {"x_astarte_realm", simple_event.realm},
-      {"x_astarte_device_id", simple_event.device_id},
-      {"x_astarte_simple_trigger_id", simple_trigger_id_str},
-      {"x_astarte_parent_trigger_id", parent_trigger_id_str},
-      {"x_astarte_event_type", to_string(event_type)}
-      | static_headers
-    ]
-
-    {routing_key, headers} =
-      update_if_has_policy(
-        routing_key,
-        headers,
-        simple_event.realm,
-        policy_name
-      )
-
-    opts_with_nil = [
-      expiration: message_expiration_ms && to_string(message_expiration_ms),
-      priority: message_priority,
-      persistent: message_persistent,
-      message_id:
-        generate_message_id(simple_event.realm, simple_event.device_id, simple_event.timestamp)
-    ]
-
-    opts = Enum.filter(opts_with_nil, fn {_k, v} -> v != nil end)
-
-    payload = SimpleEvent.encode(simple_event)
-
     result =
-      wait_ok_publish(
-        simple_event.realm,
-        target_exchange,
-        routing_key,
-        payload,
-        [{:headers, headers} | opts]
+      TriggersHandler.dispatch_event(
+        event,
+        event_type,
+        target,
+        realm,
+        device_id,
+        timestamp,
+        policy
       )
-
-    Logger.debug("headers: #{inspect(headers)}, routing key: #{inspect(routing_key)}")
 
     :telemetry.execute(
       [:astarte, :data_updater_plant, :triggers_handler, :published_event],
       %{},
       %{
-        realm: simple_event.realm,
+        realm: realm,
         event_type: to_string(event_type)
       }
     )
 
     result
-  end
-
-  defp compute_backoff_time(current_attempt) do
-    minimum_duration = (1 <<< current_attempt) * 1000
-    minimum_duration + round(minimum_duration * 0.25 * :rand.uniform())
-  end
-
-  defp generate_message_id(realm, device_id, timestamp) do
-    realm_trunc = String.slice(realm, 0..63)
-    device_id_trunc = String.slice(device_id, 0..15)
-    timestamp_hex_str = Integer.to_string(timestamp, 16)
-    rnd = Enum.random(0..@max_rand) |> Integer.to_string(16)
-
-    "#{realm_trunc}-#{device_id_trunc}-#{timestamp_hex_str}-#{rnd}"
-  end
-
-  defp update_if_has_policy(
-         "trigger_engine",
-         headers,
-         realm,
-         policy_name
-       ) do
-    policy_name = policy_name || "@default"
-
-    headers = [
-      {"x_astarte_trigger_policy", policy_name}
-      | headers
-    ]
-
-    routing_key = generate_routing_key(realm, policy_name)
-
-    {routing_key, headers}
-  end
-
-  defp update_if_has_policy(
-         routing_key,
-         headers,
-         _realm,
-         _policy_name
-       ) do
-    {routing_key, headers}
-  end
-
-  defp generate_routing_key(realm, policy) do
-    "#{realm}_#{policy}"
   end
 end
