+++ libs/astarte_events/mix.exs
@@ -66,7 +66,9 @@ defmodule Astarte.Events.Mixfile do
       # See also: https://github.com/deadtrickster/ssl_verify_fun.erl/pull/27
       {:ssl_verify_fun, "~> 1.1.0", manager: :rebar3, override: true},
       {:astarte_data_access,
-       github: "astarte-platform/astarte_data_access", branch: "release-1.3"}
+       github: "astarte-platform/astarte_data_access", branch: "release-1.3"},
+      {:astarte_core, github: "astarte-platform/astarte_core", branch: "release-1.3"},
+      {:elixir_uuid, "~> 1.2"}
     ]
   end
 end
